Finish Strategy and Factory pattern
About world : I create a WorldType class and its duty is to create and return tree and stone. When we want to add a new tree, we just need to add two classes (XStone and XTree) and add conditional statements in WorldType. Last, we should also add the new world in enum in World class.

About move step : I create a Sensibility class and its duty is to add new sensibilities rapidly. When we need to add a new sensibility, we just need to add some new conditional statement in this class and we also need to add a new statement in enum in  Player.

About unused file: the idea in this file is to create a new way to create a new world type with new class - Accessory. In the Accessory, i package up the Tree class and the Stone class. So I can just quickly create a new world class which implements this abstract class